pyproject.toml
[project]
name = "lore-analist-agent"
version = "0.1.0"
description = "AI agent for analyzing game lore, scripts, and narrative documents using RAG and web search."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.11.1",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.45.0",
    "uuid>=1.30",
]

[project.scripts]
lore-agent = "src.__main__:main"

# To run the Streamlit UI:
#   streamlit run streamlit_app.py
#   (or: streamlit run app/streamlit_app.py  after moving to app/)
